replace array<int> to array<uint8_t>

+++ 02.IP_filter/include/ip_filter.h
@@ -35,9 +35,9 @@ std::vector<std::string> split(const std::string &str, char d)
 
 
 // convert string array to uint8 array
-std::array<int, 4> convert_row(const std::vector<std::string> &row) {
+std::array<uint8_t, 4> convert_row(const std::vector<std::string> &row) {
 
-    std::array<int, 4> output;
+    std::array<uint8_t, 4> output;
     for (int i = 0; i < 4; ++i) {
         output[i] = stoi(row[i]);
     }
@@ -47,7 +47,7 @@ std::array<int, 4> convert_row(const std::vector<std::string> &row) {
 
 
 // vector comparison function
-bool vector_cmp(const std::array<int, 4> &ip1, const std::array<int, 4> &ip2) {
+bool vector_cmp(const std::array<uint8_t, 4> &ip1, const std::array<uint8_t, 4> &ip2) {
 
     if (ip1 == ip2) {
         return false;
@@ -70,7 +70,7 @@ bool vector_cmp(const std::array<int, 4> &ip1, const std::array<int, 4> &ip2) {
 
 
 // function to sort all ip in vector
-void sort_ips(std::vector<std::array<int, 4>> &ip_pool) {
+void sort_ips(std::vector<std::array<uint8_t, 4>> &ip_pool) {
 
     std::sort(ip_pool.begin(), ip_pool.end(), vector_cmp);
 
@@ -78,15 +78,15 @@ void sort_ips(std::vector<std::array<int, 4>> &ip_pool) {
 
 
 // function to filter specific byte by value
-std::vector<std::array<int, 4>> filter(
-    const std::vector<std::array<int, 4>> &ip_pool, 
+std::vector<std::array<uint8_t, 4>> filter(
+    const std::vector<std::array<uint8_t, 4>> &ip_pool, 
     int filter_byte_0 = -1, 
     int filter_byte_1 = -1, 
     int filter_byte_2 = -1, 
     int filter_byte_3 = -1
 ) {
 
-    std::vector<std::array<int, 4>> result{}; 
+    std::vector<std::array<uint8_t, 4>> result{}; 
     std::vector<int> filter_bytes{filter_byte_0, filter_byte_1, filter_byte_2, filter_byte_3};
     
     for (const auto &row : ip_pool) {
@@ -106,14 +106,14 @@ std::vector<std::array<int, 4>> filter(
 
 
 // function to filter any byte by value
-std::vector<std::array<int, 4>> filter_any(
-    const std::vector<std::array<int, 4>> &ip_pool, 
+std::vector<std::array<uint8_t, 4>> filter_any(
+    const std::vector<std::array<uint8_t, 4>> &ip_pool, 
     int filter_byte = -1
 ) {
 
-    std::vector<std::array<int, 4>> result{};
+    std::vector<std::array<uint8_t, 4>> result{};
     
-    for (std::vector<std::array<int, 4>>::const_iterator ip = ip_pool.cbegin(); ip != ip_pool.cend(); ++ip)
+    for (std::vector<std::array<uint8_t, 4>>::const_iterator ip = ip_pool.cbegin(); ip != ip_pool.cend(); ++ip)
     {
         if (std::any_of(ip->cbegin(), ip->cend(), [filter_byte](int val){ return int(val) == filter_byte; })) {
             result.emplace_back(*ip);
@@ -126,10 +126,10 @@ std::vector<std::array<int, 4>> filter_any(
 
 
 
-void print_results(const std::vector<std::array<int, 4>> &ip_pool) {
-    for (std::vector<std::array<int, 4>>::const_iterator ip = ip_pool.cbegin(); ip != ip_pool.cend(); ++ip)
+void print_results(const std::vector<std::array<uint8_t, 4>> &ip_pool) {
+    for (std::vector<std::array<uint8_t, 4>>::const_iterator ip = ip_pool.cbegin(); ip != ip_pool.cend(); ++ip)
     {
-        for (std::array<int, 4>::const_iterator ip_part = ip->cbegin(); ip_part != ip->cend(); ++ip_part)
+        for (std::array<uint8_t, 4>::const_iterator ip_part = ip->cbegin(); ip_part != ip->cend(); ++ip_part)
         {
             if (ip_part != ip->cbegin())
             {
